requirements.txt
PyYAML==6.0.3
instaloader
Gooey
